# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Admin API Documentation
  description: Dokumentasi API khusus untuk Web Admin Panel (Laravel)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /api/admin/asn:
    get:
      summary: Get All Pegawai (Filtered by Admin Organization)
      tags: [Pegawai]
      parameters:
        - in: query
          name: search
          required: false
          schema: { type: string }
          description: "Search by Nama or NIP"
      responses:
        '200':
          description: List pegawai
    post:
      summary: Create New Pegawai
      tags: [Pegawai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama: { type: string }
                nip: { type: string }
                password: { type: string }
                jabatan: { type: string }
                bidang: { type: string }
                role_id: { type: integer }
                email: { type: string, format: email }
                no_hp: { type: string }
      responses:
        '200':
          description: Pegawai created
  /api/admin/asn/{id}:
    get:
      summary: Get Pegawai Detail
      tags: [Pegawai]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Detail pegawai
    put:
      summary: Update Pegawai
      tags: [Pegawai]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nama: { type: string }
                jabatan: { type: string }
                bidang: { type: string }
                role_id: { type: integer }
                is_active: { type: boolean }
                email: { type: string, format: email }
                no_hp: { type: string }
      responses:
        '200':
          description: Pegawai updated
    delete:
      summary: Delete Pegawai
      tags: [Pegawai]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Pegawai deleted
...
  /api/admin/asn/{id}/device:
    delete:
      summary: Reset Device ID Pegawai
      tags: [Pegawai]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Device reset successful

  # --- ROLES ---
  /api/admin/roles:
    get:
      summary: Get All Roles
      tags: [Role]
      responses:
        '200':
          description: List roles
    post:
      summary: Create New Role
      tags: [Role]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_role: { type: string }
                permission_ids: 
                  type: array
                  items: { type: integer }
      responses:
        '200':
          description: Role created
...
  /api/admin/roles/{id}:
    get:
      summary: Get Role Detail
      tags: [Role]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Role detail
    put:
      summary: Update Role
      tags: [Role]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_role: { type: string }
                permission_ids: 
                  type: array
                  items: { type: integer }
      responses:
        '200':
          description: Role updated
    delete:
      summary: Delete Role
      tags: [Role]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Role deleted

  # --- JADWAL ---
...